Split the block undoing in two parts and add unit tests.
The first read the undo data from disk and the second does the undo. This ensure we don't depend on the state of the filesystem to undo and makes is easier to unit test. The patch also includes a basic unit test for undo.

(cherry picked from commit 880d847f076f7b38c849ce3f31120f2f582a0ba0)