Add a config object and pass it around.
Summary:
The code is full of global variables. We want to reduce the use of global variable. As a first step we want to introduce an API around the globals variables such as we can then swap the innards and remove the globals later on.

This is a first step in that direction. It adds a Config interface and a GlobalConfig class that map the nMaxBlockSize global variabla and then have a good chunk of the code access the global through the interface.

Test Plan:
  make check
  ../qa/pull-tester/rpc-tests.py

Reviewers: freetrader, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D29

(cherry picked from commit c0f7324b10e7b82aeb4a36fa390a5342e69b1103)

Conflicts:
	src/test/blockencodings_tests.cpp