Format baseline C/C++ files touched by ABC
The objective of this branch is to get to a reformatted baseline codebase with
minimal diffs against ABC, so that one can easier compare the changes stemming
from ABC.

NOTES:

The reformatting breaks the check-doc.py script which would need a major
overhaul to cope, and has been disabled on ABC for the moment (see D69).
For this reason, this reformatted branch does not build on Travis.