Resolve FIXMEs in src/net.cpp (pass config through)
Summary:
The config is passed through to the thread functions.
Passing by reference did not work, so pointers are used.

As there are other clients of TraceThread() which appear
non-trivial to migrate, a TraceThreadWithConfig has been
derived for wrapping those threads which currently use
the config.

Test Plan:
make check
../qa/pull-tester/rpc-tests.py

Reviewers: #bitcoin_abc!, deadalnix, sickpig, awemany, dagurval, CCulianu

Differential Revision: https://reviews.bitcoinabc.org/D311